% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.acs.demographics-package.R
\docType{package}
\name{dc.acs.demographics-package}
\alias{dc.acs.demographics}
\alias{dc.acs.demographics-package}
\title{dc.acs.demographics: Social Data Commons: Population Characteristics: ACS Demographics}
\description{
Allows user to easily get ACS demographic data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Zachary Alerte \email{rsu8zv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
